modal funtionality is done, modularization of card component and updates of css styles
