+++ frontend-test/README.md
@@ -1,70 +1,40 @@
-# Getting Started with Create React App
+# Koltin Blog Technology (Challenge Koltin México)
 
-This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).
 
-## Available Scripts
+## Índice
 
-In the project directory, you can run:
+* [1.Resumen del proyecto](#1-resumen-del-proyecto)
+* [2.Tecnologías](#2-tecnologías)
+* [3.Funcionalidad ](#3-prototipos)
 
-### `npm start`
 
-Runs the app in the development mode.\
-Open [http://localhost:3000](http://localhost:3000) to view it in your browser.
+***
 
-The page will reload when you make changes.\
-You may also see any lint errors in the console.
+## 1. Resumen del proyecto
+**Koltin Blog Technology** ¿Eres apasionado por la tecnología y te encanta estar al día con los nuevos avances del mundo Tech?
+ Koltin Blog Technology es una iniciativa de Koltin México con la que buscamos facilitarle la vida a nuestra comunidad geek, queremos ayudarte compartiendo artículos de interés y con actualizaciones continuas con solo un click. Con esta nueva alternativa podrás: 
+* Subscribirte y recibir notificaciones de actualizaciones tecnológicas.
+* Escoger la categoría de la tecnología que deseas investigar. 
+* Ver artículos recientes.
+* Dar click en una tarjeta para más información.
 
-### `npm test`
+ "Koltin Blog Technology" fue desarrollada para ser implementada en desktop. A través de tu página web "Koltin" podrás seguir disfrutando de las funcionalidades que te ofrece  con la finalidad de que puedas estar al día con noticias del mundo Tech.
 
-Launches the test runner in the interactive watch mode.\
-See the section about [running tests](https://facebook.github.io/create-react-app/docs/running-tests) for more information.
+## 2. Tecnologías.
+ 
+p align="left"> <a href="https://getbootstrap.com" target="_blank" rel="noreferrer"> <img src="https://raw.githubusercontent.com/devicons/devicon/master/icons/bootstrap/bootstrap-plain-wordmark.svg" alt="bootstrap" width="40" height="40"/> </a> <a href="https://www.w3schools.com/css/" target="_blank" rel="noreferrer"> <img src="https://raw.githubusercontent.com/devicons/devicon/master/icons/css3/css3-original-wordmark.svg" alt="css3" width="40" height="40"/> </a> <a href="https://www.figma.com/" target="_blank" rel="noreferrer"> <img src="https://www.vectorlogo.zone/logos/figma/figma-icon.svg" alt="figma" width="40" height="40"/> </a> <a href="https://git-scm.com/" target="_blank" rel="noreferrer"> <img src="https://raw.githubusercontent.com/devicons/devicon/master/icons/nodejs/nodejs-original-wordmark.svg" alt="nodejs" width="40" height="40"/> </a> <a href="https://postman.com" target="_blank" rel="noreferrer"> <img src="https://www.vectorlogo.zone/logos/getpostman/getpostman-icon.svg" alt="postman" width="40" height="40"/> </a> <a href="https://reactjs.org/" target="_blank" rel="noreferrer"> <img src="https://raw.githubusercontent.com/devicons/devicon/master/icons/react/react-original-wordmark.svg" alt="react" width="40" height="40"/> </a> <a href="https://sass-lang.com" target="_blank" rel="noreferrer">
 
-### `npm run build`
 
-Builds the app for production to the `build` folder.\
-It correctly bundles React in production mode and optimizes the build for the best performance.
+## 3. Fucionalidad.
 
-The build is minified and the filenames include the hashes.\
-Your app is ready to be deployed!
+En nuestra página web podrás disfrutar de todos los artículos actuales que traemos para ti, podrás dar click al botón 'More Article' y se mostrarán más articulos disponibles, asi como tambien darle click a la tarjeta para más información detallada.
 
-See the section about [deployment](https://facebook.github.io/create-react-app/docs/deployment) for more information.
 
-### `npm run eject`
+![pantalla1](C:\Users\rhole\Desktop\proyectos\Koltintest\Frontend-test\frontend-test\src\components\assets\Home(01))
+![pantalla2](C:\Users\rhole\Desktop\proyectos\Koltintest\Frontend-test\frontend-test\src\components\assets\BrowseCategory(02))
+![pantalla3](C:\Users\rhole\Desktop\proyectos\Koltintest\Frontend-test\frontend-test\src\components\assets\Articles(03))
+![pantalla4](C:\Users\rhole\Desktop\proyectos\Koltintest\Frontend-test\frontend-test\src\components\assets\subscribe(04))
+![pantalla5](C:\Users\rhole\Desktop\proyectos\Koltintest\Frontend-test\frontend-test\src\components\assets\Modal(05))
 
-**Note: this is a one-way operation. Once you `eject`, you can't go back!**
 
-If you aren't satisfied with the build tool and configuration choices, you can `eject` at any time. This command will remove the single build dependency from your project.
 
-Instead, it will copy all the configuration files and the transitive dependencies (webpack, Babel, ESLint, etc) right into your project so you have full control over them. All of the commands except `eject` will still work, but they will point to the copied scripts so you can tweak them. At this point you're on your own.
-
-You don't have to ever use `eject`. The curated feature set is suitable for small and middle deployments, and you shouldn't feel obligated to use this feature. However we understand that this tool wouldn't be useful if you couldn't customize it when you are ready for it.
-
-## Learn More
-
-You can learn more in the [Create React App documentation](https://facebook.github.io/create-react-app/docs/getting-started).
-
-To learn React, check out the [React documentation](https://reactjs.org/).
-
-### Code Splitting
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/code-splitting](https://facebook.github.io/create-react-app/docs/code-splitting)
-
-### Analyzing the Bundle Size
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/analyzing-the-bundle-size](https://facebook.github.io/create-react-app/docs/analyzing-the-bundle-size)
-
-### Making a Progressive Web App
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/making-a-progressive-web-app](https://facebook.github.io/create-react-app/docs/making-a-progressive-web-app)
-
-### Advanced Configuration
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/advanced-configuration](https://facebook.github.io/create-react-app/docs/advanced-configuration)
-
-### Deployment
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/deployment](https://facebook.github.io/create-react-app/docs/deployment)
-
-### `npm run build` fails to minify
-
-This section has moved here: [https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify](https://facebook.github.io/create-react-app/docs/troubleshooting#npm-run-build-fails-to-minify)
