+++ frontend-test/src/components/styles/LearnCoding.module.css
@@ -11,7 +11,7 @@ body{
   display: flex;
   flex-direction: column;
   width: auto;
-  margin: 150px 0 10px 450px;
+  margin: 150px 0 10px 550px;
 }
 
 .subscribe h2 {
