+++ frontend-test/src/components/styles/ShowArticles.module.css
@@ -22,9 +22,20 @@
     height: 425px;
     gap: 70px;
     margin: 0px 100px 50px 250px;
+    cursor: pointer;
 }
 
 .moreArticle {
     align-self: center;
     margin-bottom: 60px;
+    top:20px;
+}
+
+.modalContainer {
+    /* background-color:rgba(16, 16, 16, 0.113); */
+    position:absolute;
+    width: 100%;
+    /* height: 550px; */
+    /* /* align-items: center; */
+    
 }
